helpers in migration script
